+++ sky/shell/testing/testing.h
@@ -1,0 +1,16 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef SKY_SHELL_TESTING_TESTING_H_
+#define SKY_SHELL_TESTING_TESTING_H_
+
+namespace sky {
+namespace shell {
+
+void InitForTesting();
+
+}  // namespace shell
+}  // namespace sky
+
+#endif  // SKY_SHELL_TESTING_TESTING_H_
